last version (probably bugged)
The code seems to work at first, but then it starts throwing warnings regarding the P3P algorithm. Also, testing the program on the dataset 0 it does not work. To fix these bugs, I need to evaluate the plots of the working pipeline. Everything may be due to outliers in the KLT (I did not try a RANSAC on it), or something else.